rename search logic
